requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
sentence-transformers>=3.0.0
spacy>=3.7.5
PyMuPDF>=1.24.0
python-docx>=1.1.0
numpy>=2.1.0
scikit-learn>=1.5.0
werkzeug>=3.0.1
google-genai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0